Return  element's average orbital periods.
Return a new array that transforms the element's average altitude into their orbital periods.

The array  contain objects in the format {name: 'name', avgAlt: avgAlt}.

The values should be rounded to the nearest whole number. The body being orbited is Earth.

The radius of the earth is 6367.4447 kilometers, and the GM value of earth is 398600.4418 km3s-2.